requirements.txt
PyYAML~=6.0.1
paramiko~=3.3.1